run env var check in telemetry test file
